ARCH-B-4: narrow tool-system boundaries
