+++ src/mindroom/custom_tools/matrix_message.py
@@ -40,7 +40,10 @@ from mindroom.matrix.client_delivery import (
 from mindroom.matrix.client_thread_history import RoomThreadsPageError, get_room_threads_page
 from mindroom.matrix.mentions import format_message_with_mentions
 from mindroom.matrix.message_content import extract_and_resolve_message
-from mindroom.tool_system.runtime_context import ToolRuntimeContext, get_tool_runtime_context
+from mindroom.tool_system.runtime_context import (
+    ToolRuntimeContext,
+    get_tool_runtime_context,
+)
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
