+++ src/mindroom/custom_tools/subagents.py
@@ -22,7 +22,10 @@ from mindroom.thread_summary import (
 )
 from mindroom.thread_tags import ThreadTagsError, normalize_tag_name, set_thread_tag
 from mindroom.thread_utils import create_session_id
-from mindroom.tool_system.runtime_context import ToolRuntimeContext, get_tool_runtime_context
+from mindroom.tool_system.runtime_context import (
+    ToolRuntimeContext,
+    get_tool_runtime_context,
+)
 
 if TYPE_CHECKING:
     from pathlib import Path
