+++ src/mindroom/mcp/registry.py
@@ -4,9 +4,11 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, cast
 
-from mindroom.mcp.toolkit import MindRoomMCPToolkit, require_mcp_server_manager
-from mindroom.tool_system.metadata import (
-    _TOOL_REGISTRY,
+from mindroom.mcp.toolkit import (
+    MindRoomMCPToolkit,
+    require_mcp_server_manager,
+)
+from mindroom.tool_system.catalog import (
     TOOL_METADATA,
     ConfigField,
     SetupType,
@@ -15,6 +17,7 @@ from mindroom.tool_system.metadata import (
     ToolMetadata,
     ToolStatus,
 )
+from mindroom.tool_system.registry_state import _TOOL_REGISTRY
 
 if TYPE_CHECKING:
     from collections.abc import Callable
