+++ src/mindroom/tool_system/__init__.py
@@ -1,1 +1,1 @@
-"""Tool infrastructure: metadata, dependencies, events, plugins, skills, and sandbox proxy."""
+"""Tool-system package for MindRoom."""
