+++ src/mindroom/tool_system/metadata.py
@@ -6,19 +6,33 @@ import functools
 import importlib
 import os
 import sys
-import threading
-from contextlib import contextmanager
 from dataclasses import asdict, dataclass
 from enum import Enum
 from importlib import util as importlib_util
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Literal, cast
 
+import mindroom.tool_system.plugin_imports as plugin_module
 from mindroom.credentials import get_runtime_credentials_manager, load_scoped_credentials
 from mindroom.logging_config import get_logger
-from mindroom.tool_system import plugins as plugin_module
-from mindroom.tool_system.dependencies import auto_install_tool_extra, check_deps_installed
-from mindroom.tool_system.plugins import load_plugins
+from mindroom.tool_system.dependencies import (
+    auto_install_tool_extra,
+    check_deps_installed,
+)
+from mindroom.tool_system.registry_state import (
+    _BUILTIN_TOOL_METADATA,
+    _BUILTIN_TOOL_REGISTRY,
+    _PLUGIN_MODULE_PREFIX,
+    _PLUGIN_REGISTRATION_SCOPE,
+    _TOOL_REGISTRY,
+    TOOL_METADATA,
+    ToolMetadataValidationError,
+    _register_plugin_tool_metadata,
+    _resolved_tool_state,
+    _scoped_plugin_registration_owner,
+    _scoped_plugin_registration_store,
+    register_builtin_tool_metadata,
+)
 from mindroom.tool_system.sandbox_proxy import maybe_wrap_toolkit_for_sandbox_proxy
 from mindroom.tool_system.worker_routing import (
     ResolvedWorkerTarget,
@@ -27,7 +41,7 @@ from mindroom.tool_system.worker_routing import (
 )
 
 if TYPE_CHECKING:
-    from collections.abc import Callable, Iterator, Mapping
+    from collections.abc import Callable, Mapping
     from types import ModuleType
 
     from agno.tools import Toolkit
@@ -36,20 +50,12 @@ if TYPE_CHECKING:
     from mindroom.constants import RuntimePaths
     from mindroom.credentials import CredentialsManager
 
-# Registry mapping tool names to their factory functions
 logger = get_logger(__name__)
 
-_TOOL_REGISTRY: dict[str, Callable[[], type[Toolkit]]] = {}
-_BUILTIN_TOOL_REGISTRY: dict[str, Callable[[], type[Toolkit]]] = {}
-_PLUGIN_TOOL_METADATA_BY_MODULE: dict[str, dict[str, ToolMetadata]] = {}
-_BUILTIN_TOOL_METADATA: dict[str, ToolMetadata] = {}
 _SAFE_TOOL_INIT_OVERRIDE_FIELDS = frozenset({"base_dir", "shell_path_prepend"})
 _TEXT_CONFIG_FIELD_TYPES = frozenset({"password", "select", "text", "url"})
 AUTHORED_OVERRIDE_INHERIT = "__MINDROOM_INHERIT__"
-_PLUGIN_MODULE_PREFIX = "mindroom_plugin_"
-_TOOL_REGISTRY_STATE_LOCK = threading.RLock()
 _VALIDATION_PLUGIN_MODULE_SUFFIX = "__validation__"
-_PLUGIN_REGISTRATION_SCOPE = threading.local()
 
 
 class ToolInitOverrideError(ValueError):
@@ -60,10 +66,6 @@ class ToolConfigOverrideError(ValueError):
     """Raised when authored tool config overrides are invalid."""
 
 
-class ToolMetadataValidationError(ValueError):
-    """Raised when runtime tool metadata derived from authored config is invalid."""
-
-
 class ToolAuthoredOverrideValidator(str, Enum):
     """Explicit authored-override validation modes for a tool."""
 
@@ -71,153 +73,9 @@ class ToolAuthoredOverrideValidator(str, Enum):
     MCP = "mcp"
 
 
-@dataclass(frozen=True)
-class _ToolRegistrySnapshot:
-    registry: dict[str, Callable[[], type[Toolkit]]]
-    metadata: dict[str, ToolMetadata]
-    builtin_registry: dict[str, Callable[[], type[Toolkit]]]
-    builtin_metadata: dict[str, ToolMetadata]
-    module_import_cache: dict[Path, plugin_module._ModuleCacheEntry]
-    plugin_tool_metadata_by_module: dict[str, dict[str, ToolMetadata]]
-    plugin_modules: dict[str, ModuleType]
-
-
 def is_authored_override_inherit(value: object) -> bool:
     """Return whether an authored override value clears an inherited higher-level override."""
     return value == AUTHORED_OVERRIDE_INHERIT
-
-
-def clear_plugin_tool_registrations(module_name: str) -> None:
-    """Forget cached tool registrations for one plugin module before it is re-executed."""
-    _PLUGIN_TOOL_METADATA_BY_MODULE.pop(module_name, None)
-
-
-def snapshot_plugin_tool_registrations(module_name: str) -> dict[str, ToolMetadata]:
-    """Return a copy of one plugin module's cached tool registrations."""
-    return _PLUGIN_TOOL_METADATA_BY_MODULE.get(module_name, {}).copy()
-
-
-def restore_plugin_tool_registrations(module_name: str, registrations: dict[str, ToolMetadata]) -> None:
-    """Restore one plugin module's cached tool registrations after a failed reload."""
-    if registrations:
-        _PLUGIN_TOOL_METADATA_BY_MODULE[module_name] = registrations.copy()
-    else:
-        _PLUGIN_TOOL_METADATA_BY_MODULE.pop(module_name, None)
-
-
-@contextmanager
-def _scoped_plugin_registration_store(
-    registrations_by_module: dict[str, dict[str, ToolMetadata]],
-) -> Iterator[None]:
-    """Route plugin registration decorators into one temporary module->metadata store."""
-    sentinel = object()
-    previous = getattr(_PLUGIN_REGISTRATION_SCOPE, "registrations_by_module", sentinel)
-    _PLUGIN_REGISTRATION_SCOPE.registrations_by_module = registrations_by_module
-    try:
-        yield
-    finally:
-        if previous is sentinel:
-            delattr(_PLUGIN_REGISTRATION_SCOPE, "registrations_by_module")
-        else:
-            _PLUGIN_REGISTRATION_SCOPE.registrations_by_module = previous
-
-
-@contextmanager
-def _scoped_plugin_registration_owner(module_name: str) -> Iterator[None]:
-    """Attribute scoped validation registrations to one synthetic plugin module."""
-    sentinel = object()
-    previous = getattr(_PLUGIN_REGISTRATION_SCOPE, "owner_module_name", sentinel)
-    _PLUGIN_REGISTRATION_SCOPE.owner_module_name = module_name
-    try:
-        yield
-    finally:
-        if previous is sentinel:
-            delattr(_PLUGIN_REGISTRATION_SCOPE, "owner_module_name")
-        else:
-            _PLUGIN_REGISTRATION_SCOPE.owner_module_name = previous
-
-
-def _plugin_registration_store() -> dict[str, dict[str, ToolMetadata]]:
-    """Return the active plugin registration sink for this thread."""
-    registrations = getattr(_PLUGIN_REGISTRATION_SCOPE, "registrations_by_module", None)
-    if registrations is None:
-        return _PLUGIN_TOOL_METADATA_BY_MODULE
-    return registrations
-
-
-@contextmanager
-def locked_tool_registry_state() -> Iterator[None]:
-    """Serialize mutations of the process-global tool and plugin registries."""
-    with _TOOL_REGISTRY_STATE_LOCK:
-        yield
-
-
-def synchronize_plugin_tools(active_plugins: list[tuple[str, str]]) -> None:
-    """Rebuild the active plugin tool overlay from cached per-module registrations."""
-    desired_registry, desired_metadata = _resolved_tool_state(
-        active_plugins,
-        _PLUGIN_TOOL_METADATA_BY_MODULE,
-    )
-    _TOOL_REGISTRY.clear()
-    _TOOL_REGISTRY.update(desired_registry)
-    TOOL_METADATA.clear()
-    TOOL_METADATA.update(desired_metadata)
-
-
-def _resolved_tool_state(
-    active_plugins: list[tuple[str, str]],
-    plugin_metadata_by_module: dict[str, dict[str, ToolMetadata]],
-) -> tuple[dict[str, Callable[[], type[Toolkit]]], dict[str, ToolMetadata]]:
-    """Build one complete tool registry state from built-ins plus active plugin overlays."""
-    desired_metadata = _BUILTIN_TOOL_METADATA.copy()
-    desired_registry = _BUILTIN_TOOL_REGISTRY.copy()
-    plugin_owner_by_tool_name: dict[str, str] = {}
-
-    for plugin_name, module_name in active_plugins:
-        for tool_name, plugin_metadata in plugin_metadata_by_module.get(module_name, {}).items():
-            existing_owner = plugin_owner_by_tool_name.get(tool_name)
-            if existing_owner is not None and existing_owner != plugin_name:
-                msg = f"Plugin tool '{tool_name}' conflicts between plugins '{existing_owner}' and '{plugin_name}'."
-                raise ToolMetadataValidationError(msg)
-            plugin_owner_by_tool_name[tool_name] = plugin_name
-            desired_metadata[tool_name] = plugin_metadata
-            factory = cast("Callable[[], type[Toolkit]] | None", plugin_metadata.factory)
-            if factory is None:
-                desired_registry.pop(tool_name, None)
-            else:
-                desired_registry[tool_name] = factory
-
-    return desired_registry, desired_metadata
-
-
-def _reject_plugin_builtin_tool_collision(tool_name: str) -> None:
-    """Fail plugin registration when it reuses a built-in tool name."""
-    if tool_name in _BUILTIN_TOOL_METADATA:
-        msg = f"Plugin tool '{tool_name}' conflicts with built-in tool '{tool_name}'."
-        raise ToolMetadataValidationError(msg)
-
-
-def register_builtin_tool_metadata(metadata: ToolMetadata) -> None:
-    """Store one built-in tool or metadata-only built-in entry in the durable registry."""
-    factory = cast("Callable[[], type[Toolkit]] | None", metadata.factory)
-    _BUILTIN_TOOL_METADATA[metadata.name] = metadata
-    TOOL_METADATA[metadata.name] = metadata
-    if factory is None:
-        _BUILTIN_TOOL_REGISTRY.pop(metadata.name, None)
-        _TOOL_REGISTRY.pop(metadata.name, None)
-    else:
-        _BUILTIN_TOOL_REGISTRY[metadata.name] = factory
-        _TOOL_REGISTRY[metadata.name] = factory
-
-
-def _register_plugin_tool_metadata(module_name: str, metadata: ToolMetadata) -> None:
-    """Store one plugin tool in the per-module overlay cache."""
-    _reject_plugin_builtin_tool_collision(metadata.name)
-    module_registrations = _plugin_registration_store().setdefault(module_name, {})
-    if metadata.name in module_registrations:
-        msg = f"Plugin tool '{metadata.name}' is registered multiple times in plugin module '{module_name}'."
-        raise ToolMetadataValidationError(msg)
-    module_registrations[metadata.name] = metadata
 
 
 def apply_authored_overrides(
@@ -805,10 +663,6 @@ class ToolMetadata:
     factory: Callable | None = None  # Factory function to create tool instance
 
 
-# Global registry for tool metadata
-TOOL_METADATA: dict[str, ToolMetadata] = {}
-
-
 def register_tool_with_metadata(
     *,
     name: str,
@@ -899,67 +753,6 @@ def register_tool_with_metadata(
         return func
 
     return decorator
-
-
-def ensure_tool_registry_loaded(
-    runtime_paths: RuntimePaths,
-    config: Config | None = None,
-) -> None:
-    """Ensure core and plugin tools are registered in the metadata registry."""
-    import mindroom.tools  # noqa: F401, PLC0415  # import here to avoid tools_metadata cycle
-
-    if config is None:
-        return
-
-    load_plugins(config, runtime_paths, set_skill_roots=False)
-    from mindroom.mcp.registry import sync_mcp_tool_registry  # noqa: PLC0415
-
-    sync_mcp_tool_registry(config)
-
-
-def _capture_tool_registry_snapshot() -> _ToolRegistrySnapshot:
-    """Capture the mutable tool/plugin registry state for transactional restoration."""
-    loaded_modules = sys.modules.copy()
-    return _ToolRegistrySnapshot(
-        registry=_TOOL_REGISTRY.copy(),
-        metadata=TOOL_METADATA.copy(),
-        builtin_registry=_BUILTIN_TOOL_REGISTRY.copy(),
-        builtin_metadata=_BUILTIN_TOOL_METADATA.copy(),
-        module_import_cache=plugin_module._MODULE_IMPORT_CACHE.copy(),
-        plugin_tool_metadata_by_module={
-            module_name: registrations.copy() for module_name, registrations in _PLUGIN_TOOL_METADATA_BY_MODULE.items()
-        },
-        plugin_modules={
-            module_name: module
-            for module_name, module in loaded_modules.items()
-            if module_name.startswith(_PLUGIN_MODULE_PREFIX)
-        },
-    )
-
-
-def _restore_tool_registry_snapshot(snapshot: _ToolRegistrySnapshot) -> None:
-    """Restore one previously captured tool/plugin registry snapshot."""
-    _TOOL_REGISTRY.clear()
-    _TOOL_REGISTRY.update(snapshot.registry)
-    TOOL_METADATA.clear()
-    TOOL_METADATA.update(snapshot.metadata)
-    _BUILTIN_TOOL_REGISTRY.clear()
-    _BUILTIN_TOOL_REGISTRY.update(snapshot.builtin_registry)
-    _BUILTIN_TOOL_METADATA.clear()
-    _BUILTIN_TOOL_METADATA.update(snapshot.builtin_metadata)
-    plugin_module._MODULE_IMPORT_CACHE.clear()
-    plugin_module._MODULE_IMPORT_CACHE.update(snapshot.module_import_cache)
-    _PLUGIN_TOOL_METADATA_BY_MODULE.clear()
-    _PLUGIN_TOOL_METADATA_BY_MODULE.update(
-        {
-            module_name: registrations.copy()
-            for module_name, registrations in snapshot.plugin_tool_metadata_by_module.items()
-        },
-    )
-    for module_name in tuple(sys.modules.copy()):
-        if module_name.startswith(_PLUGIN_MODULE_PREFIX) and module_name not in snapshot.plugin_modules:
-            sys.modules.pop(module_name, None)
-    sys.modules.update(snapshot.plugin_modules)
 
 
 def _module_origin_within_root(module: ModuleType, root: Path) -> bool:
