+++ tests/test_thread_snooze_plugin.py
@@ -14,7 +14,7 @@ from unittest.mock import AsyncMock, MagicMock, call, patch
 
 import pytest
 
-import mindroom.tool_system.plugins as plugin_module
+import mindroom.tool_system.plugin_imports as plugin_module
 from mindroom.config.agent import AgentConfig
 from mindroom.config.main import Config
 from mindroom.hooks import AgentLifecycleContext, HookRegistry, ToolAfterCallContext
