+++ tests/test_workloop_thread_scope.py
@@ -16,7 +16,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-import mindroom.tool_system.plugins as plugin_module
+import mindroom.tool_system.plugin_imports as plugin_module
 from mindroom.config.agent import AgentConfig
 from mindroom.config.main import Config
 from mindroom.config.plugin import PluginEntryConfig
